+++ CHANGELOG.md
@@ -1254,3 +1254,4 @@
 - [2026-05-18 14:48] refactor(security): extract reusable atomic widgets from security
 - [2026-05-18 14:48] refactor(security): extract reusable atomic widgets from security
 - [2026-05-18 15:24] style(forms): harmonize light and dark theme palette for forms
+- [2026-05-18 15:28] test(i18n): add unit tests for i18n business logic